Slash memory use under pypy.
The chi square testing requires a large dict indexed by
permunations. This can become huge.

It's possible to "rank" a permutation p, to map it to a
unique integer in range(len(p)!). That's as small as can
be. But ranking is slow. I added a function for that
here, but don't use it.

Instead we build an int with 4 bits per element in p.
Under pypy, that generally builds an unboxed native-size
64-bit int. That's faster and much more memory frugal
then using a boxed "bigint".

It doesn't scale, but vven at 11 elements the chi square
test is pushing half an hour to finish, and going to
12 would take over 12x longer. It's fast approaching
the limit of practicality for exhaustive testing.