Move to a different PRNG. There is only 1 version now (version 0).
LCGs with a power-of-2 modulus are very weak in the low-order bits.
As list sizes get larger, that's apparently impossible to overcome
with a 16-bit LCG. We got away with it for smaller lists because
`getrandbits()` only uses the high-order bits of each output.

So now we use a 32-bit "xorshift" generator. Its low-order bits
are also weak, but not _as_ weak as in the previous LCG, and the
high-order bits we actually use are farther "to the left" than
with a 16-bit generator. And the period is much larger too.

All permutations of all lists with no more than 11 elements are
reachable now, with no chi square warnings (each possible
permutation is seen "about equally often").