Move to much stronger PRNG.
The underlying generator is now Marsaglia's xorshift128,
a 32-bit generator with 128 bits of state.

This is much stronger than xorshift32, and enough so that
Bays-Durham shuffling no longer helps.

Testing has also been boosted, including that doing a
single shuffle across objects with consecutive seeds
also delivers the expected distribution of permutations.
Since are 2**32 seeds to test, this takes a long
time. Under pypy, about 8 hours for lists of size
2 through 11 inclusive.